+++ packages/popover/src/use-floating.tsx
@@ -338,6 +338,7 @@ export default (props: PopoverPropTypes, ctx, { refReference, refContent, refArr
     const delay = resolvePopoverDelay()[1];
     popHideTimerId = setTimeout(() => {
       popShowTimerId && clearTimeout(popShowTimerId);
+      isMouseenter = false;
       localIsShow.value = false;
     }, delay);
   };
@@ -365,31 +366,30 @@ export default (props: PopoverPropTypes, ctx, { refReference, refContent, refArr
     }
   };
 
-  const handleClickRef = () => {
-    triggerPopover();
-  };
-
   const handlePopContentMouseEnter = (e: MouseEvent) => {
-    if (props.trigger !== 'hover') {
+    if ('hover' !== props.trigger) {
       return;
     }
-
-    if (popHideTimerId) {
+    if (!isMouseenter) {
       isMouseenter = true;
-      clearTimeout(popHideTimerId);
-      popHideTimerId = undefined;
     }
-
+    // 设置setTimeout的延时为delay, 避免出现同时触发mouseenter mouseleave事件
+    const delay = resolvePopoverDelay()[1];
+    if (popHideTimerId) {
+      clearTimeout(popHideTimerId);
+      setTimeout(() => {
+        popHideTimerId = undefined;
+      }, delay);
+    }
     emitPopContentMouseEnter(e);
   };
 
   const handlePopContentMouseLeave = (e: MouseEvent) => {
     // 处理底部触发mouseleave事件，popShowTimerId有值代表处于hover状态
-    if (isMouseenter && !popShowTimerId) {
+    if (popShowTimerId && !popHideTimerId) {
       hidePopover();
-      isMouseenter = false;
-      emitPopContentMouseLeave(e);
     }
+    emitPopContentMouseLeave(e);
   };
 
   /**
@@ -410,26 +410,30 @@ export default (props: PopoverPropTypes, ctx, { refReference, refContent, refArr
 
   const resolveTriggerEvents = () => {
     const triggerEvents = {
-      hover: {
-        content: [
-          ['mouseenter', handlePopContentMouseEnter],
-          ['mouseleave', handlePopContentMouseLeave],
-        ],
-        reference: [
-          ['mouseenter', showPopover],
-          ['mouseleave', hidePopover],
-          ['focus', showPopover],
-          ['blur', hidePopover],
-        ],
-      },
-      click: [['click', handleClickRef]],
-      manual: {
-        content: [
-          ['mouseenter', emitPopContentMouseEnter],
-          ['mouseleave', emitPopContentMouseLeave],
-        ],
-        reference: [[]],
-      },
+      hover: [
+        {
+          content: [
+            ['mouseenter', handlePopContentMouseEnter],
+            ['mouseleave', handlePopContentMouseLeave],
+          ],
+          reference: [
+            ['mouseenter', showPopover],
+            ['mouseleave', hidePopover],
+            ['focus', showPopover],
+            ['blur', hidePopover],
+          ],
+        },
+      ],
+      click: [[['click', showPopover]]],
+      manual: [
+        {
+          content: [
+            ['mouseenter', emitPopContentMouseEnter],
+            ['mouseleave', emitPopContentMouseLeave],
+          ],
+          reference: [[]],
+        },
+      ],
     };
 
     return triggerEvents[props.trigger] ?? [];